refactor(core): keep main/core/ UI-free
Move session timer to main/utils/ (it's LVGL-driven, not bitcoin core) and route descriptor_validator dialogs through callbacks: confirm_cb (already in the API but unused) now carries the purpose/script-binding warning, and the duplicate-detected toast is rendered by the page layer via the new descriptor_validator_get_duplicate_id() getter.
